KAFKA-5295: Refactor TopicCreation under the util package
